Information needed for auto-resolving caches.
Inform the cache whether the instance that is being looked up is unknown or known.  That way, an auto-resolving cache can decide whether it needs to query the server to satisfy the lookup or whether it can return nil and let RK create a new instance and fill it in with the server's response.